requirements.txt
fastparquet
matplotlib
numpy
pandas
pyarrow
scikit-learn
seaborn
tqdm
colorlog
openpyxl
colorama

s3fs
mlflow
cupy-cuda12x